feat: Replace output files atomically (new option: atomicReplace). (#69)
This is useful if the written files are watched by an external program
like test watchers (i.e., karma). Without an atomic replace, these
programs may process an incomplete file. This usually results in syntax
errors being reported.

This commit introduces an atomic replacement strategy whereby the new
version of a file is written to a temporary file. When the file is
completely written, the temporary file is renamed to the output file
name, replacing the contents atomically. This strategy can be disabled
via the atomicReplace boolean option.